+++ ootrando_overworldmap_hamsda/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## x
+
+- added more logic tricks
+
 ## 2.9.4.0
 
 - changed closed forest behavior to include sequence breaking Gohma